requirements.txt
fastapi
uvicorn
pymongo
pandas
flask
python-dotenv
numpy
motor
